Move classes to wireable stubs file
