+++ tests/Unit/ValidationTest.php
@@ -4,12 +4,10 @@ namespace Tests\Unit;
 
 use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Contracts\Validation\Validator as ValidatorContract;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\ViewErrorBag;
 use Livewire\Component;
 use Livewire\Livewire;
-use Livewire\Wireable;
 
 use function PHPUnit\Framework\assertTrue;
 
@@ -472,6 +470,8 @@ class ValidationTest extends TestCase
             $this->markTestSkipped('Typed Property Initialization not supported prior to PHP 7.4');
         }
 
+        require_once __DIR__.'/WireablesCanBeSetAsPublicPropertiesStubs.php';
+
         Livewire::test(ValidatesWireableProperty::class)
             ->call('runValidation')
             ->assertHasErrors('customCollection.0.amount')
@@ -864,73 +864,5 @@ class ValidatesComputedProperty extends Component
     public function render()
     {
         return view('null-view');
-    }
-}
-
-class ValidatesWireableProperty extends Component
-{
-    public CustomWireableCollection $customCollection;
-
-    public $rules = [
-        'customCollection.*.amount' => 'required|gt:100'
-    ];
-
-    public function mount()
-    {
-        $this->customCollection = new CustomWireableCollection([
-            new CustomWireableDTO(50),
-        ]);
-    }
-
-    public function runValidation()
-    {
-        $this->validate();
-    }
-
-    public function render()
-    {
-        return view('null-view');
-    }
-}
-
-class CustomWireableCollection extends Collection implements Wireable
-{
-    public function toLivewire()
-    {
-        return $this->mapWithKeys(function($dto, $key) {
-            return [$key => $dto instanceof CustomWireableDTO ? $dto->toLivewire() : $dto];
-        })->all();
-    }
-
-    public static function fromLivewire($value)
-    {
-        return static::wrap($value)
-        ->mapWithKeys(function ($dto, $key) {
-            return [$key => CustomWireableDTO::fromLivewire($dto)];
-        });
-    }
-}
-
-class CustomWireableDTO implements Wireable
-{
-    public $amount;
-
-    public function __construct($amount)
-    {
-        $this->amount = $amount;
-    }
-
-    public function toLivewire()
-    {
-        return [
-            'amount' => $this->amount
-        ];
-    }
-
-    public static function fromLivewire($value)
-    {
-        return new static(
-            $value['amount']
-        );
     }
 }
