+++ tests/Unit/WireablesCanBeSetAsPublicPropertiesStubs.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Unit;
 
+use Illuminate\Support\Collection;
 use Livewire\Component;
 use Livewire\Wireable;
 
@@ -91,5 +92,73 @@ class ComponentWithWireablePublicProperty extends Component
     public function render()
     {
         return view('wireables');
+    }
+}
+
+class ValidatesWireableProperty extends Component
+{
+    public CustomWireableCollection $customCollection;
+
+    public $rules = [
+        'customCollection.*.amount' => 'required|gt:100'
+    ];
+
+    public function mount()
+    {
+        $this->customCollection = new CustomWireableCollection([
+            new CustomWireableDTO(50),
+        ]);
+    }
+
+    public function runValidation()
+    {
+        $this->validate();
+    }
+
+    public function render()
+    {
+        return view('null-view');
+    }
+}
+
+class CustomWireableCollection extends Collection implements Wireable
+{
+    public function toLivewire()
+    {
+        return $this->mapWithKeys(function($dto, $key) {
+            return [$key => $dto instanceof CustomWireableDTO ? $dto->toLivewire() : $dto];
+        })->all();
+    }
+
+    public static function fromLivewire($value)
+    {
+        return static::wrap($value)
+        ->mapWithKeys(function ($dto, $key) {
+            return [$key => CustomWireableDTO::fromLivewire($dto)];
+        });
+    }
+}
+
+class CustomWireableDTO implements Wireable
+{
+    public $amount;
+
+    public function __construct($amount)
+    {
+        $this->amount = $amount;
+    }
+
+    public function toLivewire()
+    {
+        return [
+            'amount' => $this->amount
+        ];
+    }
+
+    public static function fromLivewire($value)
+    {
+        return new static(
+            $value['amount']
+        );
     }
 }
